+++ variables.tf
@@ -12,13 +12,19 @@ variable "nomad_namespace" {
 variable "nomad_job_name" {
   type        = string
   description = "Nomad job name"
-  default     = "presto"
+  default     = "trino"
 }
 
 variable "local_docker_image" {
   type        = bool
   description = "Switch for nomad jobs to use artifact for image lookup"
   default     = false
+}
+
+variable "consul_docker_image" {
+  type        = string
+  description = "Consul docker image"
+  default     = "consul:1.9"
 }
 
 variable "mode" {
@@ -31,45 +37,37 @@ variable "mode" {
   }
 }
 
-# presto
-variable "shared_secret_provider" {
-  type        = string
-  description = "Provider for the shared secret"
-  default     = "user"
-  validation {
-    condition     = var.shared_secret_provider == "user" || var.shared_secret_provider == "vault"
-    error_message = "Valid provider: \"user\" or \"vault\"."
-  }
-}
-
+# trino
 variable "shared_secret_user" {
   type        = string
   description = "Shared secret provided by user"
-  default     = "asdasdsadafdsa"
+  default     = "defaulttrinosecret"
   validation {
     condition     = length(var.shared_secret_user) >= 12
     error_message = "The length of the shared secret must be 12 characters or more."
   }
 }
 
-variable "shared_secret_vault" {
+variable "vault_secret" {
   type = object({
-    vault_kv_policy_name     = string,
-    vault_kv_path            = string,
-    vault_kv_secret_key_name = string
+    use_vault_provider         = bool,
+    vault_kv_policy_name       = string,
+    vault_kv_path              = string,
+    vault_kv_field_secret_name = string
   })
   description = "Set of properties to be able fetch shared cluster secret from vault"
   default = {
-    vault_kv_policy_name     = "kv-secret"
-    vault_kv_path            = "secret/data/presto"
-    vault_kv_secret_key_name = "cluster_shared_secret"
+    use_vault_provider         = true
+    vault_kv_policy_name       = "kv-secret"
+    vault_kv_path              = "secret/data/path/to/cluster-shared-secret/trino"
+    vault_kv_field_secret_name = "cluster_shared_secret"
   }
 }
 
 variable "service_name" {
   type        = string
-  description = "Presto service name"
-  default     = "presto"
+  description = "Trino service name"
+  default     = "trino"
 }
 
 variable "coordinator" {
@@ -86,7 +84,7 @@ variable "workers" {
 
 variable "debug" {
   type        = bool
-  description = "Turn on debug logging in presto nodes"
+  description = "Turn on debug logging in trino nodes"
   default     = false
 }
 
@@ -95,53 +93,93 @@ variable "consul_http_addr" {
   description = "Address to consul, resolvable from the container. e.g. http://127.0.0.1:8500"
 }
 
-variable "memory" {
-  type        = number
-  description = "Memory allocation for presto nodes"
-  default     = 1024
-  validation {
-    condition     = var.memory >= 768
-    error_message = "Presto can not run with less than 512MB of memory. 256MB is subtracted for OS. Total must be at least 768MB."
-  }
-}
-
 variable "docker_image" {
   type        = string
-  description = "Presto docker image"
-  default     = "prestosql/presto:341"
+  description = "Trino docker image"
+  default     = "trinodb/trino:354"
 }
 
 variable "consul_connect_plugin" {
-  type = bool
-  description = "Deploy consul connect plugin for presto"
-  default = true
+  type        = bool
+  description = "Deploy consul connect plugin for trino"
+  default     = true
 }
 
 variable "consul_connect_plugin_version" {
   type        = string
-  description = "Version of the consul connect plugin for presto (on maven central) src here: https://github.com/gugalnikov/presto-consul-connect"
+  description = "Version of the consul connect plugin for trino (on maven central) src here: https://github.com/gugalnikov/trino-consul-connect"
   default     = "2.2.0"
 }
 
 variable "consul_connect_plugin_artifact_source" {
-  type = string
+  type        = string
   description = "Artifact URI source"
-  default = "https://oss.sonatype.org/service/local/repositories/releases/content/io/github/gugalnikov/presto-consul-connect"
+  default     = "https://oss.sonatype.org/service/local/repositories/releases/content/io/github/gugalnikov/trino-consul-connect"
 }
 
 variable "container_environment_variables" {
   type        = list(string)
-  description = "Presto environment variables"
+  description = "Trino environment variables"
   default     = [""]
 }
 
 variable "use_canary" {
-  type = bool
-  description = "Uses canary deployment for Presto"
-  default = false
+  type        = bool
+  description = "Uses canary deployment for Trino"
+  default     = false
 }
 
-variable "hivemetastore" {
+variable "resource" {
+  type = object({
+    cpu    = number,
+    memory = number
+  })
+  default = {
+    cpu    = 500,
+    memory = 1024
+  }
+  description = "Trino resources"
+  validation {
+    condition     = var.resource.cpu >= 500 && var.resource.memory >= 768
+    error_message = "Trino can not run with less than 300Mhz CPU and less than 512MB of memory. 256MB is subtracted for OS. Total must be at least 768MB."
+  }
+}
+
+variable "resource_proxy" {
+  type = object({
+    cpu    = number,
+    memory = number
+  })
+  default = {
+    cpu    = 200,
+    memory = 128
+  }
+  description = "Trino proxy resources"
+  validation {
+    condition     = var.resource_proxy.cpu >= 200 && var.resource_proxy.memory >= 128
+    error_message = "Proxy resource must be at least: cpu=200, memory=128."
+  }
+}
+
+variable "hive_config_properties" {
+  type        = list(string)
+  description = "Custom hive configuration properties"
+  default     = [""]
+}
+
+# ML model
+variable "model_service_name" {
+  type = string
+  description = "ML model service name"
+  default = "model"
+}
+
+######
+# Service dependencies
+######
+
+## Hive
+variable "hivemetastore_service" {
   type = object({
     service_name = string,
     port         = number,
@@ -153,7 +191,8 @@ variable "hivemetastore" {
   description = "Hivemetastore data-object contains service_name and port"
 }
 
-variable "minio" {
+## Minio
+variable "minio_service" {
   type = object({
     service_name = string,
     port         = number,
@@ -163,22 +202,49 @@ variable "minio" {
   description = "Minio data-object contains service_name, port, access_key and secret_key"
 }
 
-# presto 
-variable "presto" {
+variable "minio_vault_secret" {
   type = object({
-    service_name = string,
-    port = number,
+    use_vault_provider         = bool,
+    vault_kv_policy_name       = string,
+    vault_kv_path              = string,
+    vault_kv_field_access_name = string,
+    vault_kv_field_secret_name = string
   })
+  description = "Set of properties to be able to fetch secret from vault"
   default = {
-    service_name = "presto"
-    port         = 8080
+    use_vault_provider         = false
+    vault_kv_policy_name       = "kv-secret"
+    vault_kv_path              = "secret/data/dev/trino"
+    vault_kv_field_access_name = "access_key"
+    vault_kv_field_secret_name = "secret_key"
   }
-  description = "Presto data-object contains service_name and port"
 }
 
-# ML model
-variable "model_service_name" {
-  type        = string
-  description = "ML model service name"
-  default     = "model"
+# Postgres
+variable "postgres_service" {
+  type = object({
+    service_name  = string,
+    port          = number,
+    username      = string,
+    password      = string,
+    database_name = string,
+  })
+  description = "Postgres data-object contains service_name, port, username, password and database_name"
+}
+variable "postgres_vault_secret" {
+  type = object({
+    use_vault_provider      = bool,
+    vault_kv_policy_name    = string,
+    vault_kv_path           = string,
+    vault_kv_field_username = string,
+    vault_kv_field_password = string
+  })
+  description = "Set of properties to be able to fetch Postgres secrets from vault"
+  default = {
+    use_vault_provider      = false
+    vault_kv_policy_name    = "kv-secret"
+    vault_kv_path           = "secret/data/dev/trino"
+    vault_kv_field_username = "username"
+    vault_kv_field_password = "password"
+  }
 }